+++ src/mimic/orm/macros/src/imp/mod.rs
@@ -31,6 +31,19 @@ pub fn any<N: MacroNode>(node: &N, t: Trait) -> TokenStream {
     let def = node.def();
 
     let imp = match t {
+        Trait::NodeDyn => {
+            let q = quote! {
+                fn path_dyn(&self) -> String {
+                    <Self as ::mimic::orm::traits::Path>::path()
+                }
+                fn serialize_dyn(&self) -> Result<Vec<u8>, ::mimic::orm::Error> {
+                    ::mimic::orm::serialize(&self)
+                }
+            };
+
+            Implementor::new(def, t).set_tokens(q).to_token_stream()
+        }
+
         Trait::Path => {
             let ident_str = format!("{}", def.ident);
             let q = quote! {